adopting ansible-galaxy collection

+++ collector/tools/dev.sh
@@ -1,16 +1,13 @@
-#!/bin/bash -x
-
-project="collector-openshift"
-clear && mkdir -p /tmp/bundle
-
-sudo podman run -it --rm --pull always \
-    --name ${project} -h ${project} \
-    --volume $(pwd):/root/koffer:z \
-    --volume /tmp/bundle:/root/bundle:z \
-    --workdir /root/koffer --entrypoint bash \
-    --volume ${HOME}/.docker:/root/.docker:ro \
-    --volume ${HOME}/.gitconfig:/root/.gitconfig:z \
+#!/bin/bash
+dirPath="$(pwd)"
+project="collector-sparta"
+mkdir -p ${dirPath}/{bundle,.docker}
+touch ${dirPath}/.gitconfig
+clear && sudo podman run -it --rm --pull always \
+    --name    ${project} -h ${project}                 \
+    --volume  ${dirPath}:/root/koffer:z                \
+    --volume  ${dirPath}/bundle:/root/bundle:z         \
+    --volume  ${dirPath}/.docker:/root/.docker:ro      \
+    --volume  ${dirPath}/.gitconfig:/root/.gitconfig:z \
+    --workdir /root/koffer --entrypoint bash           \
   quay.io/cloudctl/koffer:latest
-
-#   --volume ${HOME}/.aws:/root/.aws:z \
-#   --volume ${HOME}/.docker:/root/.docker:z \
